finished everything but user timer input

+++ main.c
@@ -16,8 +16,8 @@
 #define MAX_LEN 100
 #define MAX_ARGS 100 
 #define MAX_PROGRAMS 100 
-static  struct itimerval timer;
-volatile int Current_Pid; 
+static struct itimerval timer;
+volatile pid_t Current_Pid; 
 
 
 typedef struct pid_queue
@@ -33,21 +33,28 @@ pid_queue queue = {0};
 void push( pid_queue * queue, int pid);
 int pop (pid_queue * queue); 
 
-void handle_sigusr1(int sig) {
-    printf("remver\n");
-}
+//void handle_sigusr1(int sig) {
+//    printf("remver\n");/
+//}
 
 
 void timer_handler(int signum){
+     puts("timer");
+
     if(Current_Pid !=0){
         kill(Current_Pid, SIGSTOP);
         push(&queue,Current_Pid);
-        Current_Pid = pop(&queue);
     }
+    //reinstall the singal handler
+      signal(SIGALRM, timer_handler);
 
-    setitimer(ITIMER_REAL, &timer, NULL);
+    //setitimer(ITIMER_REAL, &timer, NULL);
 
-    puts("timer");
+}
+
+void alarm_handler(int signum) {
+    // Handle the SIGALRM signal
+    printf("SIGALRM received!\n");
 }
 
 
@@ -56,19 +63,19 @@ int main(int argc, char * argv[])
     uint32_t schedule_time = argv[1];
     
     signal(SIGALRM, timer_handler);
+    //signal(SIGALARM, alarm_handler);
     // Convert milliseconds to microseconds
     int milliseconds = 1000; // 1000 milliseconds = 1 second
     int microseconds = milliseconds * 1000;
 
     // Configure the timer to expire after 'milliseconds'
     timer.it_value.tv_sec = 2;
-    timer.it_value.tv_usec = 0;
-    timer.it_interval.tv_sec = 0; // non-periodic timer
-    timer.it_interval.tv_usec = 0;
+    timer.it_value.tv_usec =0;
+    timer.it_interval.tv_sec =0; 
+    timer.it_interval.tv_usec =0;
+
 
     char *args[MAX_ARGS]; 
-    // char arguments[MAX_ARGS][MAX_LEN] = {{0},{0}}; 
-    //char arguemnts[100] = {0};
     uint8_t num_of_commands =  argc -1; 
    
     //char arg[100]; 
@@ -77,7 +84,7 @@ int main(int argc, char * argv[])
     uint8_t num_of_programs = 0; 
     uint8_t num_of_arguments = 0; 
     void * null_ptr = NULL;
-    uint8_t pid = 0; 
+    pid_t pid = 0; 
     char * colon = ":"; 
     
     char program[100] = {0}; 
@@ -105,25 +112,20 @@ int main(int argc, char * argv[])
 
         if(strcmp(argv[i], colon) == 0 || i==num_of_commands )
          {
-
+            
             pid = fork();
-        
+           
             if(pid == 0)
             {
-                //struct sigaction sa = {0}; 
-                //sa.sa_flags = SIGCONT;
-                //sa.sa_handler = &handle_sigusr1;
-                //sigaction(SIGUSR1, &sa,NULL);
-                //pause();
-                kill(getpid(), SIGSTOP);
+                raise(SIGSTOP);
                 execv(args[0], args);     
             }
-
+             printf("child pid %d\n",pid);
             program_pid[pid_index] = pid; 
             push(&queue,pid);
             pid_index++;
             num_of_programs++; 
-            next_program = 0; 
+            next_program = 0;
             p_index = 0;
         } 
 
@@ -134,38 +136,41 @@ int main(int argc, char * argv[])
 
 
    
-    while(queue.head || queue.tail)
+    while(queue.head !=-1 && queue.tail !=-1)
     {   
         int status; 
         //printf("program pid %d\n", program_pid[i]);
         Current_Pid = pop(&queue);
         //printf("pid %d",id);
+        if(Current_Pid != 0)
+        {
           if (kill(Current_Pid, SIGCONT) == -1){
+            printf("curr %d\n", Current_Pid);
+            printf("fukced up\n");
+
+          }
+        } 
+      
+       
+       if (setitimer(ITIMER_REAL, &timer, NULL) == -1){
+        printf("timer ufkced up\n");
+
+
 
        }
 
-         if (setitimer(ITIMER_REAL, &timer, NULL) == -1) 
-        {
-            perror("setitimer");
-            exit(EXIT_FAILURE);
-        }
-     
-     
-
-      
-
-      
-       // waitpid(Current_Pid, &status, WUNTRACED);
+       waitpid(Current_Pid, &status, WUNTRACED);
 
    }  
 
-    wait(NULL);
+   wait(NULL);
    printf("finished\n"); 
 }
 
     
 void push( pid_queue * queue, int pid)
 {
+    
     queue->pids[queue->tail] = pid; 
     queue->tail++; 
 }
@@ -173,11 +178,12 @@ void push( pid_queue * queue, int pid)
 int pop (pid_queue * queue)
 {
     int pid = 0; 
+   
     pid = queue->pids[queue->head];
     queue->head +=1;
    if (queue->head > queue->tail) {// Reset queue when no elements are left
-            queue->head = 0;
-            queue->tail= 0; 
+            queue->head = -1;
+            queue->tail= -1; 
    } 
    return pid;
 }
